update contributor service and controller

+++ src/modules/expense-sharing/controllers/expense-group-members.controller.ts
@@ -5,6 +5,7 @@ import { ExpenseGroupMemberMapper } from '../mappers/expense-group-member.mapper
 import { UpdateExpenseGroupMemberDto } from '../dto/requests/update-expense-group-member-dto';
 import { CreateExpenseGroupMemberDto } from '../dto/requests/create-expense-group-member-dto';
 import { ExpenseGroupService } from '../services/expense-group.service';
+import { User } from 'src/common/decorators/user.decorator';
 
 @Controller('expense-groups/:groupId/members')
 @UseGuards(AuthGuard('jwt'))
@@ -37,12 +38,16 @@ export class ExpenseGroupMembersController {
 
     @Patch(':memberId')
     async updateMember(
+        @User('userId') userId: string,
         @Param('groupId') groupId: string,
         @Param('memberId') memberId: string,
         @Body(ValidationPipe) input: UpdateExpenseGroupMemberDto,
     ) {
         const group = await this.groupService.findOne(groupId);
         if (!group) throw new HttpException('Group not found', HttpStatus.NOT_FOUND);
+
+        const isAdmin = await this.memberService.isAdmin(group, userId);
+        if (!isAdmin) throw new HttpException('You do not have permission to update members in this group', HttpStatus.UNAUTHORIZED);
 
         const updatedMember = await this.memberService.updateMember(memberId, input);
         if (!updatedMember) throw new HttpException('Updated member not found', HttpStatus.NOT_FOUND);
@@ -51,9 +56,12 @@ export class ExpenseGroupMembersController {
     }
 
     @Delete(':memberId')
-    async removeMember(@Param('groupId') groupId: string, @Param('memberId') memberId: string) {
+    async removeMember(@User('userId') userId: string, @Param('groupId') groupId: string, @Param('memberId') memberId: string) {
         const group = await this.groupService.findOne(groupId);
         if (!group) throw new HttpException('Group not found', HttpStatus.NOT_FOUND);
+
+        const isAdmin = await this.memberService.isAdmin(group, userId);
+        if (!isAdmin) throw new HttpException('You do not have permission to remove members from this group', HttpStatus.UNAUTHORIZED);
 
         await this.memberService.removeMember(group.id, memberId);
     }
